Move page/grading into page/academy/grading
