+++ src/page/academy/grading/subcomponents/GradingEditorComponent.tsx
@@ -12,9 +12,9 @@ import { IconNames } from '@blueprintjs/icons';
 import * as React from 'react';
 import ReactMde, { ReactMdeProps } from 'react-mde';
 import { Prompt } from 'react-router';
-import controlButton from '../../../commons/ControlButton';
-import Markdown from '../../../commons/Markdown';
 
+import controlButton from 'src/commons/ControlButton';
+import Markdown from 'src/commons/Markdown';
 import { getPrettyDate } from 'src/utils/dateHelpers';
 import { showSuccessMessage, showWarningMessage } from 'src/utils/notification';
 import { stringParamToInt } from 'src/utils/paramParseHelpers';
